Syncronized changes with bitbucket repository.

+++ Test_Server/server.h
@@ -21,7 +21,7 @@ private:
     QSslSocket *socket;
     DBConn *dbConn;
 
-    int _dataSize;
+    qint64 _dataSize;
     bool _readHeader;
     quint32 _readType;
